Revert "Changed module name to library. Added permission checks. Getting events from one calendar working."
This reverts commit 01b21b1f
