+++ app/src/main/java/net/planner/planet/User.kt
@@ -1,0 +1,10 @@
+package net.planner.planet
+
+class User {
+    // Gets user accounts from Android AccountManager
+    //@TODO might need to use OAuth2
+
+    // https://developer.android.com/training/id-auth/identify
+    // https://developer.android.com/reference/android/accounts/AccountManager
+
+}
