+++ app/src/test/java/net/planner/planet/ExampleUnitTest.kt
@@ -14,10 +14,4 @@ class ExampleUnitTest {
     fun addition_isCorrect() {
         assertEquals(4, 2 + 2)
     }
-
-    @Test
-    fun testGetCalender() {
-        val communicator = GoogleCalenderCommunicator()
-
-    }
 }
